// src/types/enum.ts
export enum UserRole {
    USER = "user",
    ADMIN = "admin",
}
export enum UserGender {
    MALE = "male",
    FEMALE = "female",
}
export enum AddressLabel {
    HOME = "home",
    WORK = "work",
    OTHER = "other",
}
export enum OrderStatus {
    PENDING = "pending",
    PREPARING = "preparing",
    SHIPPED = "shipped",
    CANCELLED_BY_USER = "cancelledByUser",
    CANCELLED_BY_ADMIN = "cancelledByAdmin",
    REFUSED = "refused",
    RECEIVED = "received",
}
export enum TestimonialStatus {
    PENDING = "pending",
    APPROVED = "approved",
    REFUSED = "refused",
}